Split `sr-api-macros` into multiple files
